Huge User and Server interface refactor.
When comparing the old interface to RFC5054 as well as trying to use the library in
practice, it became clear that the pysrp interface that was used as
inspiration was not designed with the typical SRP workflow. Namely,
the server should provide the SRP primitives prior to registration and
login, which means the server will 'go first' rather than the user.
This is contrary to the design of pysrp which assumes the challenge
from the server comes later.

Besides leading to a wonky interface that doesn't quite match the
actual usage pattern, it created a very tricky internal User object state
that required juggling a bunch of late-initialized variables. Shooting
oneself in the foot was much easier by trying to access things like
the session key prior to calling processChallenge. Now there is no
worry of late-initialized public variables.

The new interface combines User object creation, startAuthentication
and processChallenge into a single public User factory method which
takes user creds and the server-provided challenge. The challenge now
also includes all the SRP primitives.

createSaltedVerificationKey during registration is now static so no
need to create a User object then. This avoids the awkwardness of
handling two scenarios of creating a User object: registration and
authentication, which require different variables and have different uses.

While this refactor emphasizes the server providing the SRP
primitives, there is nothing preventing the user from providing the
primitives by creating a Challenge object with them.